+++ CHANGELOG.md
@@ -4,3 +4,5 @@
     [JENKINS-23603](https://issues.jenkins-ci.org/browse/JENKINS-23603)
   * set up the build to use http://repo.jenkins-ci.org/public/ similar to the Maven POMs
     [JENKINS-19942](https://issues.jenkins-ci.org/browse/JENKINS-19942)
+  * register input files for the localizer task to avoid that the task is skipped when the inputs change
+    [JENKINS-24298](https://issues.jenkins-ci.org/browse/JENKINS-24298)